refactor: replace getAllMdList with getAllCloudMdList for improved TOC file handling
